+++ sbin/dataimport-04-fix-data-in-production-database.sh
@@ -67,7 +67,7 @@ select yn in "Yes"; do
     esac
 done
 
-echo "YOUR TODO: start object-queue-receiver in a separate shell 'cd ${MYTB} && node dist/backend/serverAdmin.js -c ${CONFIG_BASEDIR}backend.dev.json --command objectDetectionManager --action receiveQueueResponses --debug 1 && cd $CWD"
+echo "YOUR TODO: start object-queue-receiver in a separate shell 'cd ${MYTB} && node dist/backend/serverAdmin.js --adminclibackend ${CONFIG_BASEDIR}adminCli.dev.json -c ${CONFIG_BASEDIR}backend.dev.json --command objectDetectionManager --action receiveQueueResponses --debug 1 && cd $CWD"
 echo "OPEN: Can we proceed the next steps?"
 select yn in "Yes"; do
     case $yn in
@@ -77,8 +77,8 @@ done
 
 echo "now: send to queue (2 runs -> to get picasa-files blocked till picasa finished)"
 cd ${MYTB}
-node dist/backend/serverAdmin.js -c ${CONFIG_BASEDIR}backend.dev.json --command objectDetectionManager --action sendQueueRequests --detector picasafile --maxPerRun 2000 --debug 1
-node dist/backend/serverAdmin.js -c ${CONFIG_BASEDIR}backend.dev.json --command objectDetectionManager --action sendQueueRequests --detector tfjs_cocossd_mobilenet_v1,tfjs_cocossd_mobilenet_v2,tfjs_cocossd_lite_mobilenet_v2,tfjs_mobilenet_v1,faceapi --maxPerRun 2000 --debug 1
+node dist/backend/serverAdmin.js --adminclibackend ${CONFIG_BASEDIR}adminCli.dev.json -c ${CONFIG_BASEDIR}backend.dev.json --command objectDetectionManager --action sendQueueRequests --detector picasafile --maxPerRun 2000 --debug 1
+node dist/backend/serverAdmin.js --adminclibackend ${CONFIG_BASEDIR}adminCli.dev.json -c ${CONFIG_BASEDIR}backend.dev.json --command objectDetectionManager --action sendQueueRequests --detector tfjs_cocossd_mobilenet_v1,tfjs_cocossd_mobilenet_v2,tfjs_cocossd_lite_mobilenet_v2,tfjs_mobilenet_v1,faceapi --maxPerRun 2000 --debug 1
 cd $CWD
 
 echo "done - fix data in production-database"