+++ src/backend/serverAdmin.ts
@@ -1,7 +1,16 @@
 import minimist from 'minimist';
-import {AdminCommandManager} from './commands/admin-command.manager';
+import {AdminCommandConfigType, AdminCommandManager} from './commands/admin-command.manager';
+import fs from 'fs';
 
 const argv = minimist(process.argv.slice(2));
+
+const filePathConfigJson = argv['adminclibackend'];
+if (filePathConfigJson === undefined) {
+    console.error('ERROR - parameters required adminclibackend: "--adminclibackend"');
+    process.exit(-1);
+}
+
+const adminBackendConfig: AdminCommandConfigType = JSON.parse(fs.readFileSync(filePathConfigJson, { encoding: 'utf8' }));
 
 // disable debug-logging
 const debug = argv['debug'] || false;
@@ -13,7 +22,7 @@ if (!debug || debug === false || parseInt(debug, 10) < 1) {
     console.debug = function() {};
 }
 
-const adminCommandManager = new AdminCommandManager();
+const adminCommandManager = new AdminCommandManager(adminBackendConfig);
 adminCommandManager.process(argv).then(value => {
     console.log('DONE - command finished:', value, argv);
     process.exit(0);